% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{ci_variable}
\alias{ci_variable}
\title{gets the confidence interval and p-value for a single variable}
\usage{
ci_variable(base_est, base_se, resampled_ests, resampled_ses, probs)
}
\arguments{
\item{base_est}{base model estimate}

\item{base_se}{base model SE}

\item{resampled_ests}{Vector of estimates from resampling}

\item{resampled_ses}{Vector of standard errors from resampling}
}
\value{
Returns a length 3 vector: left and right CI value, and p-value
}
\description{
gets the confidence interval and p-value for a single variable
}
\keyword{internal}
