% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{detect_cores}
\alias{detect_cores}
\title{Calculate number of cores using parallel, or not if parallel not loaded}
\usage{
detect_cores()
}
\description{
Calculate number of cores using parallel, or not if parallel not loaded
}
\keyword{internal}
