% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootGlmm.R
\name{DetectCores}
\alias{DetectCores}
\title{Calculate number of cores using parallel,
or not if parallel not loaded}
\usage{
DetectCores()
}
\description{
Calculate number of cores using parallel,
or not if parallel not loaded
}
\keyword{internal}
