% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{find_bars}
\alias{find_bars}
\title{Returns the terms with bars from a formula}
\usage{
find_bars(form_with_bars)
}
\arguments{
\item{form_with_bars}{A formula used in e.g. \code{lme4} and similar
packages. Typically along the lines:
\code{y ~ age + (1 | school)} etc}
}
\description{
Returns the terms with bars from a formula
}
\keyword{internal}
