% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{FindBars}
\alias{FindBars}
\title{returns the terms with bars}
\usage{
FindBars(form_with_bars)
}
\arguments{
\item{form_with_bars}{A formula used in e.g. lme4 and similar
packages. Typically along the lines:
y ~ age + (1 | school)
etc}
}
\description{
returns the terms with bars
}
\keyword{internal}
