% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{gen_resampling_index}
\alias{gen_resampling_index}
\title{Given resampled vectors, gives matching index of original variables}
\usage{
gen_resampling_index(orig_list, sampled_list)
}
\arguments{
\item{orig_list}{List of original data vectors.}

\item{sampled_list}{Sampled list}
}
\value{
Returns an index vector.
}
\description{
This function takes in original vectors, and resampled editions,
it spits back the matching index of the original variables
for the new resampled ones.
}
\keyword{internal}
