% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{gen_samp_lev}
\alias{gen_samp_lev}
\title{For resampling from a single set of levels}
\usage{
gen_samp_lev(levels, unique_lim = NULL, reduce_by_one = TRUE)
}
\arguments{
\item{levels}{The levels to sample}

\item{unique_lim}{Default NULL; optionally a required number of unique
elements to have in the sample}

\item{reduce_by_one}{Logical, default TRUE; for (potentially) more accurate coverage,
resample one less than the number of levels in each list}
}
\value{
A sample of 'levels'; a vector.
}
\description{
For resampling from a single set of levels
}
\keyword{internal}
