% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{gen_sample}
\alias{gen_sample}
\title{Generates block resamples}
\usage{
gen_sample(list_of_levels, rand_columns = NULL,
  unique_resample_lim = NULL, reduce_by_one = TRUE)
}
\arguments{
\item{list_of_levels}{list of vectors of levels of each random effect}

\item{rand_columns}{Default NULL.
name of columns to randomise over; if NULL, will use
all in 'list_of_levels'}

\item{unique_resample_lim}{Default NULL; optionally set a minimal number of unique levels each sample
must produce. Note that it should be a named vector,
the same as the levels to randomise over.}

\item{reduce_by_one}{Logical, default TRUE; for (potentially) more accurate coverage,
resample one less than the number of levels in each list}
}
\value{
A list of samples
}
\description{
Takes in a list of unique levels in the random columns,
gives back a random sampling of each.
}
\keyword{internal}
