% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{GenResamplingIndex}
\alias{GenResamplingIndex}
\title{this function takes in original vectors, and resampled editions,
it spits back the matching index of the original variables for the new resampled ones}
\usage{
GenResamplingIndex(orig_list, sampled_list)
}
\description{
this function takes in original vectors, and resampled editions,
it spits back the matching index of the original variables for the new resampled ones
}
\keyword{internal}
