% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{GenSample}
\alias{GenSample}
\title{Generates block resamples}
\usage{
GenSample(list_of_levels, rand_columns = NULL,
  unique_resample_lim = NULL, reduce_by_one = TRUE)
}
\description{
Takes in a list of unique levels in the random columns,
gives back a random sampling of each.
}
\keyword{internal}
