% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{list_of_matrices}
\alias{list_of_matrices}
\title{Checks that an object is a list of matrices}
\usage{
list_of_matrices(list_to_check, allow_null = TRUE)
}
\arguments{
\item{list_to_check}{The "list" (maybe!) to check}

\item{allow_null}{If an element is NULL, is that OK?}
}
\value{
TRUE or FALSE
}
\description{
Checks that an object is a list, and also
that the list is a collection of matrices.
Currently returns FALSE on an empty list
}
\keyword{internal}
