% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_model.R
\name{bootstrap_runner}
\alias{bootstrap_runner}
\title{Runs the bootstrapping of the models}
\usage{
bootstrap_runner(bootstrap_function, resamples, parallelism = c("none",
  "future", "parallel"), num_cores = NULL)
}
\arguments{
\item{bootstrap_function}{a function that we wish to run `resamples` times}

\item{resamples}{how many times we wish to run `bootstrap_function`}

\item{parallelism}{How to run this function in parallel}

\item{num_cores}{How many cores to use, if using `parallel`}
}
\value{
returns the list that contains the results of running
  `bootstrap_function`.
}
\description{
This function gets passed a function that runs a single bootstrap resample
and a number of resamples, and decides how to run them
e.g. in parallel etc
}
\keyword{internal}
